requirements.txt
anyio==3.6.2
chess==1.9.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-dotenv==0.21.1
rfc3986==1.5.0
sniffio==1.3.0
stockfish==3.28.0
